+++ src/main/java/chronometry/HexaghostModel.java
@@ -1,63 +1,19 @@
 package chronometry;
 
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Graphics;
-import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.math.MathUtils;
-import com.megacrit.cardcrawl.actions.AbstractGameAction.AttackEffect;
-import com.megacrit.cardcrawl.actions.GameActionManager;
-import com.megacrit.cardcrawl.actions.animations.AnimateSlowAttackAction;
-import com.megacrit.cardcrawl.actions.animations.VFXAction;
-import com.megacrit.cardcrawl.actions.common.ApplyPowerAction;
-import com.megacrit.cardcrawl.actions.common.ChangeStateAction;
-import com.megacrit.cardcrawl.actions.common.DamageAction;
-import com.megacrit.cardcrawl.actions.common.GainBlockAction;
-import com.megacrit.cardcrawl.actions.common.MakeTempCardInDiscardAction;
-import com.megacrit.cardcrawl.actions.common.RollMoveAction;
-import com.megacrit.cardcrawl.actions.unique.BurnIncreaseAction;
-import com.megacrit.cardcrawl.actions.utility.SFXAction;
-import com.megacrit.cardcrawl.audio.MusicMaster;
-import com.megacrit.cardcrawl.audio.SoundMaster;
-import com.megacrit.cardcrawl.cards.DamageInfo;
-import com.megacrit.cardcrawl.cards.status.Burn;
-import com.megacrit.cardcrawl.characters.AbstractPlayer;
-import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.megacrit.cardcrawl.core.Settings;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
-import com.megacrit.cardcrawl.helpers.Hitbox;
-import com.megacrit.cardcrawl.helpers.ScreenShake;
-import com.megacrit.cardcrawl.localization.LocalizedStrings;
-import com.megacrit.cardcrawl.localization.MonsterStrings;
-import com.megacrit.cardcrawl.monsters.AbstractMonster;
-import com.megacrit.cardcrawl.monsters.AbstractMonster.EnemyType;
-import com.megacrit.cardcrawl.monsters.AbstractMonster.Intent;
-import com.megacrit.cardcrawl.powers.StrengthPower;
-import com.megacrit.cardcrawl.rooms.AbstractRoom;
-import com.megacrit.cardcrawl.rooms.MonsterRoomBoss;
-import com.megacrit.cardcrawl.scenes.AbstractScene;
-import com.megacrit.cardcrawl.unlock.UnlockTracker;
-import com.megacrit.cardcrawl.vfx.BorderFlashEffect;
-import com.megacrit.cardcrawl.vfx.combat.FireballEffect;
-import com.megacrit.cardcrawl.vfx.combat.GhostIgniteEffect;
-import com.megacrit.cardcrawl.vfx.combat.InflameEffect;
-import com.megacrit.cardcrawl.vfx.combat.ScreenOnFireEffect;
-import com.megacrit.cardcrawl.vfx.BobEffect;
-import com.megacrit.cardcrawl.vfx.TintEffect;
-import com.megacrit.cardcrawl.helpers.MathHelper;
-
 import com.megacrit.cardcrawl.helpers.ImageMaster;
-import java.util.ArrayList;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import chronometry.SlayTheStreamer;
+import com.megacrit.cardcrawl.helpers.MathHelper;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import com.megacrit.cardcrawl.vfx.BobEffect;
 
 public class HexaghostModel extends AbstractMonster {
   public static final String ID = "Hexaghost";
   public static final String IMAGE = "images/monsters/theBottom/boss/ghost/core.png";
-  
+
   private Texture plasma1;
   private Texture plasma2;
   private Texture plasma3;
@@ -75,7 +31,7 @@ public class HexaghostModel extends AbstractMonster {
   public HexaghostModel()
   {
     super("Hexaghost", "Hexaghost", 250, 20.0F, 0.0F, 450.0F, 450.0F, "images/monsters/theBottom/boss/ghost/core.png");
-    this.type = AbstractMonster.EnemyType.BOSS;
+    this.type = EnemyType.BOSS;
 
     this.plasma1 = ImageMaster.loadImage("images/monsters/theBottom/boss/ghost/plasma1.png");
     this.plasma2 = ImageMaster.loadImage("images/monsters/theBottom/boss/ghost/plasma2.png");
@@ -133,13 +89,13 @@ public class HexaghostModel extends AbstractMonster {
     // Render Body
     sb.setColor(this.tint.color);
     sb.draw(this.plasma3, this.drawX - 256.0F + this.animX + 12.0F * Settings.scale,
-            this.drawY + this.animY + AbstractDungeon.sceneOffsetY + this.effect.y * 2.0F - 256.0F * Settings.scale + BODY_OFFSET_Y * Settings.scale, 256.0F, 256.0F, 512.0F, 512.0F, Settings.scale * 0.475F, Settings.scale * 0.475F, this.plasma3Angle, 0, 0, 512, 512, false, false);
+            this.drawY + this.animY + AbstractDungeon.sceneOffsetY + this.effect.y * 2.0F - 256.0F * Settings.scale + BODY_OFFSET_Y, 256.0F, 256.0F, 512.0F, 512.0F, Settings.scale * 0.475F, Settings.scale * 0.475F, this.plasma3Angle, 0, 0, 512, 512, false, false);
     sb.draw(this.plasma2, this.drawX - 256.0F + this.animX + 6.0F * Settings.scale,
-            this.drawY + this.animY + AbstractDungeon.sceneOffsetY + this.effect.y - 256.0F * Settings.scale + BODY_OFFSET_Y * Settings.scale, 256.0F, 256.0F, 512.0F, 512.0F, Settings.scale*0.5F, Settings.scale*0.5F, this.plasma2Angle, 0, 0, 512, 512, false, false);
+            this.drawY + this.animY + AbstractDungeon.sceneOffsetY + this.effect.y - 256.0F * Settings.scale + BODY_OFFSET_Y, 256.0F, 256.0F, 512.0F, 512.0F, Settings.scale*0.5F, Settings.scale*0.5F, this.plasma2Angle, 0, 0, 512, 512, false, false);
     sb.draw(this.plasma1, this.drawX - 256.0F + this.animX,
-            this.drawY + this.animY + AbstractDungeon.sceneOffsetY + this.effect.y * 0.5F - 256.0F * Settings.scale + BODY_OFFSET_Y * Settings.scale, 256.0F, 256.0F, 512.0F, 512.0F, Settings.scale*0.5F, Settings.scale*0.5F, this.plasma1Angle, 0, 0, 512, 512, false, false);
+            this.drawY + this.animY + AbstractDungeon.sceneOffsetY + this.effect.y * 0.5F - 256.0F * Settings.scale + BODY_OFFSET_Y, 256.0F, 256.0F, 512.0F, 512.0F, Settings.scale*0.5F, Settings.scale*0.5F, this.plasma1Angle, 0, 0, 512, 512, false, false);
     sb.draw(this.shadow, this.drawX - 256.0F + this.animX + 12.0F * Settings.scale,
-            this.drawY + this.animY + AbstractDungeon.sceneOffsetY + this.effect.y / 4.0F - 15.0F * Settings.scale - 256.0F * Settings.scale + BODY_OFFSET_Y * Settings.scale, 256.0F, 256.0F, 512.0F, 512.0F, Settings.scale*0.5F, Settings.scale*0.5F, 0.0F, 0, 0, 512, 512, false, false);
+            this.drawY + this.animY + AbstractDungeon.sceneOffsetY + this.effect.y / 4.0F - 15.0F * Settings.scale - 256.0F * Settings.scale + BODY_OFFSET_Y, 256.0F, 256.0F, 512.0F, 512.0F, Settings.scale*0.5F, Settings.scale*0.5F, 0.0F, 0, 0, 512, 512, false, false);
 
     // Render Core
     sb.draw(this.core,
