+++ src/main/java/chronometry/MonsterMessageRepeater.java
@@ -1,39 +1,9 @@
 package chronometry;
 
-import com.megacrit.cardcrawl.localization.*;
-import com.megacrit.cardcrawl.monsters.exordium.*;
-import com.megacrit.cardcrawl.monsters.city.*;
-import com.megacrit.cardcrawl.monsters.beyond.*;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
-import com.megacrit.cardcrawl.screens.mainMenu.*;
 import com.megacrit.cardcrawl.dungeons.*;
-import com.megacrit.cardcrawl.characters.*;
-import com.megacrit.cardcrawl.helpers.controller.*;
 import com.megacrit.cardcrawl.core.*;
 import com.megacrit.cardcrawl.actions.animations.*;
-import com.megacrit.cardcrawl.rooms.*;
-import com.megacrit.cardcrawl.vfx.*;
-import com.megacrit.cardcrawl.helpers.*;
-import com.megacrit.cardcrawl.unlock.*;
-import com.megacrit.cardcrawl.vfx.GlowyFireEyesEffect;
-import com.megacrit.cardcrawl.vfx.StaffFireEffect;
-import com.megacrit.cardcrawl.vfx.combat.InflameEffect;
-
-import com.badlogic.gdx.*;
-import com.badlogic.gdx.math.*;
-import com.badlogic.gdx.graphics.g2d.*;
-import com.badlogic.gdx.graphics.*;
-import com.esotericsoftware.spine.Skeleton;
-
-import java.util.*;
-import java.util.stream.*;
-import java.util.function.*;
-import java.lang.reflect.*;
-import org.apache.logging.log4j.*;
-import de.robojumper.ststwitch.*;
-
-import chronometry.SlayTheStreamer;
-import basemod.ReflectionHacks;
 
 public class MonsterMessageRepeater {
