+++ src/main/java/chronometry/patches/NoSkipBossRelicPatch.java
@@ -4,54 +4,18 @@ import java.util.ArrayList;
 
 import com.evacipated.cardcrawl.modthespire.lib.SpireReturn;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
-import com.evacipated.cardcrawl.modthespire.lib.SpireInsertPatch;
-import com.evacipated.cardcrawl.modthespire.lib.SpireEnum;
-import com.evacipated.cardcrawl.modthespire.lib.ByRef;
 import com.megacrit.cardcrawl.core.Settings;
-import com.megacrit.cardcrawl.core.AbstractCreature;
-import com.megacrit.cardcrawl.core.CardCrawlGame;
-import com.megacrit.cardcrawl.neow.NeowEvent;
-import com.megacrit.cardcrawl.cards.AbstractCard;
-import com.megacrit.cardcrawl.cards.CardGroup;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
-import com.megacrit.cardcrawl.unlock.UnlockTracker;
-import com.megacrit.cardcrawl.helpers.ImageMaster;
-import com.megacrit.cardcrawl.map.DungeonMap;
-import com.megacrit.cardcrawl.map.MapRoomNode;
-import com.megacrit.cardcrawl.rooms.AbstractRoom;
-import com.megacrit.cardcrawl.rooms.TreasureRoom;
-import com.megacrit.cardcrawl.ui.buttons.ProceedButton;
-import com.megacrit.cardcrawl.helpers.Hitbox;
-import com.megacrit.cardcrawl.helpers.controller.CInputAction;
-import com.megacrit.cardcrawl.helpers.controller.CInputActionSet;
 import com.megacrit.cardcrawl.screens.select.BossRelicSelectScreen;
 
-import com.megacrit.cardcrawl.localization.*;
 import com.megacrit.cardcrawl.relics.*;
-import com.megacrit.cardcrawl.blights.*;
-import com.megacrit.cardcrawl.screens.mainMenu.*;
-import com.badlogic.gdx.*;
-import com.megacrit.cardcrawl.dungeons.*;
-import com.badlogic.gdx.math.*;
-import com.megacrit.cardcrawl.characters.*;
-import com.megacrit.cardcrawl.helpers.controller.*;
-import com.megacrit.cardcrawl.core.*;
-import com.megacrit.cardcrawl.rooms.*;
-import com.badlogic.gdx.graphics.g2d.*;
-import com.megacrit.cardcrawl.vfx.*;
 import com.badlogic.gdx.graphics.*;
 import com.megacrit.cardcrawl.helpers.*;
-import com.megacrit.cardcrawl.unlock.*;
 import java.util.*;
-import java.util.stream.*;
-import java.util.function.*;
-import org.apache.logging.log4j.*;
 import de.robojumper.ststwitch.*;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import chronometry.SlayTheStreamer;
-import chronometry.BossSelectRoom;
-import basemod.ReflectionHacks;
 
 public class NoSkipBossRelicPatch {
